server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
pyjwt==2.8.0
certifi==2023.7.22
requests==2.31.0
discord.py==2.3.2
py-trello==0.19.0 